requirements.txt
aiofiles==22.1.0
alembic==1.9.2
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==7.1.2
colorama==0.4.6
cookiecutter==1.7.3
decorator==4.4.2
dnspython==2.3.0
email-validator==1.3.0
fastapi==0.68.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.0
hyperframe==5.2.0
idna==2.10
imageio==2.25.0
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
jinja2-time==0.2.0
Mako==1.2.4
manage-fastapi==1.1.1
MarkupSafe==2.1.2
more-itertools==9.0.0
moviepy==1.0.3
mysqlclient==2.1.1
numpy==1.24.1
openai-whisper==20230124
packaging==23.0
Pillow==9.4.0
poyo==0.5.0
proglog==0.1.10
prompt-toolkit==3.0.36
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-slugify==7.0.0
PyYAML==6.0
qrcode==7.3.1
questionary==1.10.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.14.2
text-unidecode==1.3
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.12.2
wcwidth==0.2.6
